+++ core/agent_harness/prompts/assistant.py
@@ -2,12 +2,10 @@
 
 from __future__ import annotations
 
-import json
-import logging
-from pathlib import Path
-from typing import TYPE_CHECKING, Any, Protocol
+from typing import TYPE_CHECKING, Protocol
 
 from config.constants.prompts import SUGGESTED_PROMPT_AFTER_FAILED_SYNTHETIC_TEST
+from core.agent_harness.prompts import synthetic_failure
 from core.agent_harness.prompts.assistant_agent_prompt import (
     _build_observation_block,
     _build_system_prompt,
@@ -18,18 +16,14 @@ from core.agent_harness.prompts.conversation_memory import (
     format_recent_conversation,
 )
 from core.agent_harness.prompts.prior_investigation import (
-    STALE_PRIOR_INVESTIGATION_NOTE,
+    build_block as build_prior_investigation_block,
+)
+from core.agent_harness.prompts.prior_investigation import (
     is_prior_investigation_follow_up,
-    is_within_recall_window,
-    prior_investigation_headline,
 )
 
 if TYPE_CHECKING:
     from core.agent_harness.turns.turn_snapshot import TurnSnapshot
-
-_logger = logging.getLogger(__name__)
-
-_MAX_SYNTHETIC_OBSERVATION_PROMPT_CHARS = 120_000
 
 
 class AssistantPromptContextProvider(Protocol):
@@ -91,96 +85,6 @@ def build_observation_block(tool_observation: str | None, *, on_screen: bool = T
     return _build_observation_block(tool_observation, on_screen=on_screen)
 
 
-def _summarize_evidence(evidence: Any) -> list[str]:
-    if isinstance(evidence, dict):
-        sample_keys = list(evidence)[:3]
-        sample = {key: evidence[key] for key in sample_keys}
-        return [
-            f"Evidence items: {len(evidence)}",
-            "Evidence keys: " + ", ".join(map(str, sample_keys)),
-            "Sample evidence:\n" + json.dumps(sample, indent=2, default=str)[:1500],
-        ]
-    if isinstance(evidence, list):
-        return [
-            f"Evidence items: {len(evidence)}",
-            "Sample evidence:\n" + json.dumps(evidence[:3], indent=2, default=str)[:1500],
-        ]
-    return [
-        f"Evidence type: {type(evidence).__name__}",
-        f"Evidence summary:\n{str(evidence)[:1500]}",
-    ]
-
-
-def _prior_investigation_block(
-    state: dict[str, Any] | None,
-    *,
-    explicit_follow_up: bool = False,
-) -> str:
-    """Summarize the session's investigation, marking it background when incidental.
-
-    An explicit follow-up handoff means the user asked about *this* investigation,
-    so it is the answer — never downgrade it to background there, however long ago
-    it ran. The note applies only when the block is unprompted context on some
-    other turn, where fresh evidence should win a conflict.
-    """
-    if state is None:
-        return ""
-    summary = _summarize_last_state(state)
-    if explicit_follow_up or is_within_recall_window(state):
-        return summary
-    return f"{STALE_PRIOR_INVESTIGATION_NOTE}\n{summary}"
-
-
-def _summarize_last_state(state: dict[str, Any]) -> str:
-    """Produce a compact text summary of the previous investigation."""
-    parts: list[str] = prior_investigation_headline(state)
-    problem_md = state.get("problem_md") or ""
-    if problem_md:
-        parts.append(f"Problem summary:\n{problem_md[:2000]}")
-    slack_message = state.get("slack_message") or ""
-    if slack_message:
-        parts.append(f"Report:\n{slack_message[:2000]}")
-    evidence = state.get("evidence")
-    if evidence:
-        try:
-            parts.extend(_summarize_evidence(evidence))
-        except (TypeError, ValueError) as exc:
-            _logger.warning("could not serialize evidence for grounding: %s", exc)
-            parts.append("(evidence present but could not be serialized for grounding)")
-    return "\n\n".join(parts) or "(no prior investigation details available)"
-
-
-def _user_message_requests_synthetic_failure_explanation(
-    message: str,
-    suggested_prompt: str = SUGGESTED_PROMPT_AFTER_FAILED_SYNTHETIC_TEST,
-) -> bool:
-    """True when the user is likely asking about a failed synthetic benchmark."""
-    m = message.strip().lower()
-    if not m:
-        return False
-    suggested = suggested_prompt.lower().rstrip("?")
-    if m.rstrip("?") == suggested:
-        return True
-    if "why" in m and "fail" in m:
-        return True
-    return "what went wrong" in m
-
-
-def _load_synthetic_observation_text(
-    path_str: str, *, max_chars: int = _MAX_SYNTHETIC_OBSERVATION_PROMPT_CHARS
-) -> str:
-    try:
-        raw = Path(path_str).read_text(encoding="utf-8")
-    except OSError:
-        return ""
-    if len(raw) > max_chars:
-        return (
-            raw[:max_chars]
-            + f"\n… [truncated for prompt size; observation is {len(raw)} characters total]"
-        )
-    return raw
-
-
 def _assistant_context_blocks(
     *,
     turn_snapshot: TurnSnapshot,
@@ -193,7 +97,7 @@ def _assistant_context_blocks(
         f"{_build_integration_guard(turn_snapshot)}"
         f"{build_handoff_guidance_block(handoff_contents)}"
         f"{build_observation_block(tool_observation, on_screen=tool_observation_on_screen)}"
-        f"{_build_synthetic_failure_block(turn_snapshot, suggested_prompt=suggested_prompt)}"
+        f"{synthetic_failure.build_block(turn_snapshot, suggested_prompt=suggested_prompt)}"
     )
 
 
@@ -208,35 +112,6 @@ def _build_integration_guard(ctx: TurnSnapshot) -> str:
         "`/mcp connect <server>` for MCP servers. Do not claim any integration is "
         "already connected, and for show/verify/remove requests against unconfigured "
         "integrations, answer with guidance only.\n\n"
-    )
-
-
-def _build_synthetic_failure_block(
-    ctx: TurnSnapshot,
-    *,
-    suggested_prompt: str = SUGGESTED_PROMPT_AFTER_FAILED_SYNTHETIC_TEST,
-) -> str:
-    obs_path = ctx.last_synthetic_observation_path
-    if not obs_path:
-        return ""
-
-    if not _user_message_requests_synthetic_failure_explanation(
-        ctx.text,
-        suggested_prompt=suggested_prompt,
-    ):
-        return ""
-
-    obs_text = _load_synthetic_observation_text(obs_path)
-    if not obs_text:
-        return ""
-
-    return (
-        "The user is asking about a failed `opensre tests synthetic` run "
-        "in this checkout. The JSON below is the saved observation "
-        f"(scores, gates, stderr summary). Path: {obs_path}\n"
-        "Use it to explain validation failures. Do not say nothing ran or "
-        "that you lack context — the run completed and this file was written.\n\n"
-        f"--- observation_json ---\n{obs_text}\n\n"
     )
 
 
@@ -262,7 +137,7 @@ def build_cli_agent_prompt_from_provider(
         # it would make OpenSRE claim it lacks incident details it still holds.
         # Age only downgrades it to background (see the note), because past the
         # recall window the turn also gathers fresh evidence.
-        prior_investigation=_prior_investigation_block(
+        prior_investigation=build_prior_investigation_block(
             turn_snapshot.last_state,
             explicit_follow_up=is_prior_investigation_follow_up(handoff_contents),
         ),
