+++ core/agent_harness/prompts/prior_investigation.py
@@ -9,9 +9,13 @@ one bound.
 
 from __future__ import annotations
 
+import json
+import logging
 import time
 from collections.abc import Mapping
 from typing import Any
+
+_logger = logging.getLogger(__name__)
 
 # Used only when the turn is *not* an explicit follow-up handoff: age softens
 # incidental prior-RCA context (stale note on the answer prompt; gather prompt
@@ -52,6 +56,65 @@ def prior_investigation_headline(state: Mapping[str, Any]) -> list[str]:
     return parts
 
 
+def _summarize_evidence(evidence: Any) -> list[str]:
+    if isinstance(evidence, dict):
+        sample_keys = list(evidence)[:3]
+        sample = {key: evidence[key] for key in sample_keys}
+        return [
+            f"Evidence items: {len(evidence)}",
+            "Evidence keys: " + ", ".join(map(str, sample_keys)),
+            "Sample evidence:\n" + json.dumps(sample, indent=2, default=str)[:1500],
+        ]
+    if isinstance(evidence, list):
+        return [
+            f"Evidence items: {len(evidence)}",
+            "Sample evidence:\n" + json.dumps(evidence[:3], indent=2, default=str)[:1500],
+        ]
+    return [
+        f"Evidence type: {type(evidence).__name__}",
+        f"Evidence summary:\n{str(evidence)[:1500]}",
+    ]
+
+
+def summarize(state: Mapping[str, Any]) -> str:
+    """Produce a compact text summary of the previous investigation."""
+    parts: list[str] = prior_investigation_headline(state)
+    problem_md = state.get("problem_md") or ""
+    if problem_md:
+        parts.append(f"Problem summary:\n{problem_md[:2000]}")
+    slack_message = state.get("slack_message") or ""
+    if slack_message:
+        parts.append(f"Report:\n{slack_message[:2000]}")
+    evidence = state.get("evidence")
+    if evidence:
+        try:
+            parts.extend(_summarize_evidence(evidence))
+        except (TypeError, ValueError) as exc:
+            _logger.warning("could not serialize evidence for grounding: %s", exc)
+            parts.append("(evidence present but could not be serialized for grounding)")
+    return "\n\n".join(parts) or "(no prior investigation details available)"
+
+
+def build_block(
+    state: Mapping[str, Any] | None,
+    *,
+    explicit_follow_up: bool = False,
+) -> str:
+    """Summarize the session's investigation, marking it background when incidental.
+
+    An explicit follow-up handoff means the user asked about *this* investigation,
+    so it is the answer — never downgrade it to background there, however long ago
+    it ran. The note applies only when the block is unprompted context on some
+    other turn, where fresh evidence should win a conflict.
+    """
+    if state is None:
+        return ""
+    summary = summarize(state)
+    if explicit_follow_up or is_within_recall_window(state):
+        return summary
+    return f"{STALE_PRIOR_INVESTIGATION_NOTE}\n{summary}"
+
+
 def is_within_recall_window(state: Mapping[str, Any] | None) -> bool:
     """True when ``state`` is recent enough to answer from without gathering.
 
@@ -77,7 +140,9 @@ __all__ = [
     "PRIOR_INVESTIGATION_HANDOFF_PREFIX",
     "PRIOR_INVESTIGATION_RECALL_SECONDS",
     "STALE_PRIOR_INVESTIGATION_NOTE",
+    "build_block",
     "is_prior_investigation_follow_up",
     "is_within_recall_window",
     "prior_investigation_headline",
+    "summarize",
 ]
